fix(compiler): move stack analysis utilities
Move known integer and stack address metadata helpers under stackAnalysis/utils, and update stack analysis imports directly without compatibility exports.
