+++ packages/compiler/src/stackAnalysis/utils/knownIntegerValue.ts
